+++ examples/__init__.py
@@ -1,0 +1,1 @@
+"""Palm example apps and definition packs (not part of the installable library)."""